refactor(draug): A.4a — move refactor_types to draug-daemon
Proof-of-concept for the Phase A.4 file-by-file migration pattern.

* Adds draug-daemon as a path dep on compositor (will be removed
  in A.5 once compositor stops using draug-daemon types directly)
* Moves the RefactorTask + TaskStatus types from
  compositor/src/refactor_types.rs to draug-daemon/src/refactor_types.rs
* Compositor file becomes a `pub use draug_daemon::refactor_types::*`
  shim so existing call sites resolve unchanged

Active-log update documents the dependency audit done while
planning the rest of A.4. Net finding: ~5500 of the 7000 LOC are
cleanly movable; ~1500 (token_stream + parts of agent_logic,
autodream) stay in compositor because they own UI state.
