+++ userspace/draug-daemon/src/lib.rs
@@ -18,3 +18,5 @@
 #![no_std]
 
 extern crate alloc;
+
+pub mod refactor_types;
